requirements.txt
networkx==2.8.8
pickleDB==0.9.2
tweepy==4.12.1
matplotlib==3.6.2
streamlit==1.15.2
numpy==1.23.5
scipy==1.9.3
pandas==1.5.2
python-dotenv==0.21.0